+++ lib/Braintree/Dispute.php
@@ -37,6 +37,11 @@ class Dispute extends Base
     const STANDARD        = 'standard';
     const NOT_PROTECTED   = 'not_protected';
 
+    /* Dispute ProtectionLevel */
+    const EFFORTLESS_CBP   = 'Effortless Chargeback Protection tool';
+    const STANDARD_CBP     = 'Chargeback Protection tool';
+    const NO_PROTECTION    = 'No Protection';
+
     /* Dispute Kind */
     const CHARGEBACK      = 'chargeback';
     const PRE_ARBITRATION = 'pre_arbitration';
@@ -45,6 +50,13 @@ class Dispute extends Base
     protected function _initialize($disputeAttribs)
     {
         $this->_attributes = $disputeAttribs;
+
+        if (isset($disputeAttribs['chargebackProtectionLevel']) && in_array($disputeAttribs['chargebackProtectionLevel'], array(self::EFFORTLESS, self::STANDARD))) {
+            $protectionLevel = constant('self::' . strtoupper($disputeAttribs['chargebackProtectionLevel']) . '_CBP');
+            $this->_set('protectionLevel', $protectionLevel);
+        } else {
+            $this->_set('protectionLevel', self::NO_PROTECTION);
+        }
 
         if (isset($disputeAttribs['transaction'])) {
             $transactionDetails = new Dispute\TransactionDetails($disputeAttribs['transaction']);
@@ -210,8 +222,11 @@ class Dispute extends Base
         return Configuration::gateway()->dispute()->search($query);
     }
 
-    /*
+    // NEXT_MAJOR_VERSION Remove this function
+    /**
      * Retrive all types of chargeback protection level types
+     *
+     * @deprecated Use allProtectionLevelTypes() instead
      *
      * @return array
      */
@@ -221,6 +236,20 @@ class Dispute extends Base
             Dispute::EFFORTLESS,
             Dispute::STANDARD,
             Dispute::NOT_PROTECTED
+        ];
+    }
+
+   /*
+     * Retrieve all types of protection level types
+     *
+     * @return array
+     */
+    public static function allProtectionLevelTypes()
+    {
+        return [
+            Dispute::EFFORTLESS_CBP,
+            Dispute::STANDARD_CBP,
+            Dispute::NO_PROTECTION
         ];
     }
 }